prep refactoring for solana package
