+++ dev.html
@@ -7,7 +7,7 @@
     <title>Development</title>
     <script src="https://cdn.ethers.io/lib/ethers-5.7.umd.min.js" type="application/javascript"></script>
     <script crossorigin src="https://unpkg.com/@depay/solana-web3.js@1"></script>
-    <script crossorigin src="https://unpkg.com/@depay/web3-blockchains@7"></script>
+    <script crossorigin src="https://unpkg.com/@depay/web3-blockchains@8"></script>
     <script crossorigin src="https://unpkg.com/@depay/walletconnect-v1@1"></script>
     <script crossorigin src="https://unpkg.com/@depay/web3-wallets@12"></script>
     <script crossorigin src="https://unpkg.com/@depay/web3-client@10"></script>
