+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@depay/web3-payments",
   "moduleName": "Web3Payments",
-  "version": "11.1.0",
+  "version": "11.2.0",
   "description": "JavaScript library to scan crypto wallets for liquefiable assets and perform cost-effective, auto-converted payments on-chain.",
   "main": "dist/umd/index.js",
   "module": "dist/esm/index.js",
@@ -10,7 +10,7 @@
     "dist"
   ],
   "scripts": {
-    "build": "rm -rf dist && rollup -c rollup.module.config.js && rollup -c rollup.module.evm.config.js",
+    "build": "rm -rf dist && rollup -c rollup.module.config.js && rollup -c rollup.module.evm.config.js && rollup -c rollup.module.solana.config.js",
     "lint": "eslint \"src/**/*.{js,jsx,ts,tsx}\"",
     "lint:fix": "eslint \"src/**/*.{js,jsx,ts,tsx}\" --fix",
     "prepublishOnly": "yarn build",
@@ -36,10 +36,11 @@
   "private": false,
   "dependencies": {},
   "peerDependencies": {
-    "@depay/web3-assets": "^6.7.0",
-    "@depay/web3-blockchains": "^7.1.0",
-    "@depay/web3-exchanges": "^11.6.1",
-    "@depay/web3-tokens": "^9.16.0",
+    "@depay/solana-web3.js": "^1.21.2",
+    "@depay/web3-assets": "^6.8.0",
+    "@depay/web3-blockchains": "^8.0.1",
+    "@depay/web3-exchanges": "^12.1.0",
+    "@depay/web3-tokens": "^9.18.0",
     "ethers": "^5.7.1"
   },
   "engines": {
@@ -49,25 +50,26 @@
     "@babel/core": "^7.12.9",
     "@babel/preset-env": "^7.12.7",
     "@depay/coinbase-wallet-sdk": "^3.6.4",
-    "@depay/solana-web3.js": "^1.20.15",
+    "@depay/solana-web3.js": "^1.21.2",
     "@depay/walletconnect-v1": "^1.8.0",
     "@depay/walletconnect-v2": "^2.4.2",
-    "@depay/web3-assets": "^6.7.0",
-    "@depay/web3-assets-evm": "^6.7.0",
-    "@depay/web3-blockchains": "^7.1.0",
-    "@depay/web3-client": "^10.7.1",
-    "@depay/web3-client-evm": "^10.7.1",
-    "@depay/web3-client-solana": "^10.7.1",
-    "@depay/web3-exchanges": "^11.6.1",
-    "@depay/web3-exchanges-evm": "^11.6.1",
-    "@depay/web3-exchanges-solana": "^11.6.1",
-    "@depay/web3-mock": "^13.26.0",
-    "@depay/web3-tokens": "^9.16.0",
-    "@depay/web3-tokens-evm": "^9.16.0",
-    "@depay/web3-tokens-solana": "^9.16.0",
-    "@depay/web3-wallets": "^15.4.0",
-    "@depay/web3-wallets-evm": "^15.4.0",
-    "@depay/web3-wallets-solana": "^15.4.0",
+    "@depay/web3-assets": "^6.8.0",
+    "@depay/web3-assets-evm": "^6.8.0",
+    "@depay/web3-assets-solana": "^6.8.0",
+    "@depay/web3-blockchains": "^8.0.1",
+    "@depay/web3-client": "^10.11.0",
+    "@depay/web3-client-evm": "^10.11.0",
+    "@depay/web3-client-solana": "^10.11.0",
+    "@depay/web3-exchanges": "^12.1.0",
+    "@depay/web3-exchanges-evm": "^12.1.0",
+    "@depay/web3-exchanges-solana": "^12.1.0",
+    "@depay/web3-mock": "^14.7.0",
+    "@depay/web3-tokens": "^9.18.0",
+    "@depay/web3-tokens-evm": "^9.18.0",
+    "@depay/web3-tokens-solana": "^9.18.0",
+    "@depay/web3-wallets": "^15.6.0",
+    "@depay/web3-wallets-evm": "^15.6.0",
+    "@depay/web3-wallets-solana": "^15.6.0",
     "@rollup/plugin-commonjs": "^18.0.0",
     "@rollup/plugin-node-resolve": "^11.2.1",
     "@rollup/plugin-replace": "^2.4.2",
@@ -75,6 +77,7 @@
     "@types/node": "^18.15.9",
     "babel-loader": "^8.2.2",
     "babel-preset-env": "^1.7.0",
+    "decimal.js": "^10.4.3",
     "eslint": "^7.15.0",
     "eslint-loader": "^4.0.2",
     "eslint-plugin-import": "^2.22.1",
