+++ package.solana.json
@@ -1,10 +1,10 @@
 {
-  "name": "@depay/web3-payments-evm",
+  "name": "@depay/web3-payments-solana",
   "moduleName": "Web3Payments",
   "version": "11.2.0",
   "description": "JavaScript library to scan crypto wallets for liquefiable assets and perform cost-effective, auto-converted payments on-chain.",
-  "main": "dist/umd/index.evm.js",
-  "module": "dist/esm/index.evm.js",
+  "main": "dist/umd/index.solana.js",
+  "module": "dist/esm/index.solana.js",
   "files": [
     "dist"
   ],
@@ -26,10 +26,11 @@
   "private": false,
   "dependencies": {},
   "peerDependencies": {
-    "@depay/web3-assets-evm": "^6.8.0",
+    "@depay/solana-web3.js": "^1.21.2",
+    "@depay/web3-assets-solana": "^6.8.0",
     "@depay/web3-blockchains": "^8.0.1",
-    "@depay/web3-exchanges-evm": "^12.1.0",
-    "@depay/web3-tokens-evm": "^9.18.0",
+    "@depay/web3-exchanges-solana": "^12.1.0",
+    "@depay/web3-tokens-solana": "^9.18.0",
     "ethers": "^5.7.1"
   },
   "engines": {
