+++ src/platforms/solana/routers.js
@@ -1,0 +1,6 @@
+export default {
+  solana: {
+    address: '0xae60aC8e69414C2Dc362D0e6a03af643d1D85b92',
+    api: []
+  },
+}
