+++ src/platforms/solana/transaction.js
@@ -1,0 +1,16 @@
+/*#if _EVM
+
+
+/*#elif _SOLANA
+
+//#else */
+
+//#endif
+
+let getTransaction = async({ paymentRoute, event, fee })=> {
+  return
+}
+
+export {
+  getTransaction
+}
